Reorganize code a bit
